refactor(web): organize codebase
