+++ apps/web/src/app/profile/page.tsx
@@ -1,19 +1,3 @@
-import { redirect } from "next/navigation";
-import { headers } from "next/headers";
-import { authClient } from "@/lib/auth-client";
-import { UserProfile } from "@/components/user-profile";
+import ProfilePage from "@/features/profile/routes/profile-page";
 
-export default async function ProfilePage() {
-	const session = await authClient.getSession({
-		fetchOptions: {
-			headers: await headers(),
-			throw: true,
-		},
-	});
-
-	if (!session?.user) {
-		redirect("/login");
-	}
-
-	return <UserProfile />;
-}
+export default ProfilePage;
