+++ apps/web/src/components/providers.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import { ThemeProvider } from "./theme-provider";
-import { Toaster } from "./ui/sonner";
+import { ThemeProvider } from "@/components/theme-provider";
+import { Toaster } from "@/components/ui/sonner";
 
 export default function Providers({ children }: { children: React.ReactNode }) {
   return (
