+++ apps/web/src/features/auth/routes/login-page.tsx
@@ -1,0 +1,15 @@
+"use client";
+
+import SignInForm from "@/features/auth/components/sign-in-form";
+import SignUpForm from "@/features/auth/components/sign-up-form";
+import { useState } from "react";
+
+export default function LoginPage() {
+	const [showSignIn, setShowSignIn] = useState(false);
+
+	return showSignIn ? (
+		<SignInForm onSwitchToSignUp={() => setShowSignIn(false)} />
+	) : (
+		<SignUpForm onSwitchToSignIn={() => setShowSignIn(true)} />
+	);
+}
